Refactor data creation and restructure code
