+++ ui-toolkit/src/main/kotlin/io/snabble/sdk/shopfinder/ShopListFragment.kt
@@ -8,6 +8,7 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.Fragment
 import io.snabble.sdk.Snabble
 import io.snabble.sdk.checkin.CheckInLocationManager
+import io.snabble.sdk.shopfinder.shoplist.ExpandableShopListRecyclerView
 import io.snabble.sdk.ui.toolkit.R
 import io.snabble.sdk.utils.isNotNullOrBlank
