+++ ui-toolkit/src/main/kotlin/io/snabble/sdk/shopfinder/shoplist/ViewType.kt
@@ -1,0 +1,11 @@
+package io.snabble.sdk.shopfinder.shoplist
+
+enum class ViewType(val isShop: Boolean) {
+    CollapsedBrand(false),
+    ExpandedBrand(false),
+    HiddenShop(true),
+    VisibleShop(true);
+
+    val isBrand: Boolean
+        get() = !isShop
+}
